moved Link account actions to new file called LinkTokenActions.js. Removed actions from index.js
